Soldan** client connected to cloud. Working version

+++ src/main/java/main/converter/ThresholdToDataSendCoverterImpl.java
@@ -16,10 +16,10 @@ public class ThresholdToDataSendCoverterImpl implements IThresholdToDataSendCove
 
 
 
-        String hex = String.format("%04X",(int)threshold.getMaxTemp()) +
-                String.format("%04X", (int)threshold.getMinTemp()) +
-                String.format("%04X", (int)threshold.getMaxHumidity()) +
-                String.format("%04X", (int)threshold.getMinHumidity());
+        String hex = String.format("%04X",threshold.getMaxTemp()) +
+                String.format("%04X", threshold.getMinTemp()) +
+                String.format("%04X", threshold.getMaxHumidity()) +
+                String.format("%04X", threshold.getMinHumidity());
 
         DataSend dataToSend = new DataSend("tx", threshold.getRoomId(), 2, false, hex);
         return dataToSend;
